organizing debug files
