+++ src/debug.c
@@ -6,7 +6,7 @@
 /*   By: mely-pan <mely-pan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/31 19:28:47 by mely-pan          #+#    #+#             */
-/*   Updated: 2025/05/31 19:36:32 by mely-pan         ###   ########.fr       */
+/*   Updated: 2025/06/02 19:50:26 by mely-pan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
