+++ src/execute/redir.c
@@ -6,7 +6,7 @@
 /*   By: mely-pan <mely-pan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/19 20:34:14 by mely-pan          #+#    #+#             */
-/*   Updated: 2025/05/31 22:06:16 by mely-pan         ###   ########.fr       */
+/*   Updated: 2025/06/02 20:04:37 by mely-pan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
